pyproject.toml
[project]
name = "flow-local"
version = "0.1.0"
description = "Fully local dictation for macOS Apple Silicon"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Isaac Horowitz" }]

dependencies = [
    "mlx>=0.18.0",
    "mlx-lm>=0.19.0",
    "parakeet-mlx>=0.5.1",
    "numpy>=1.26",
    "sounddevice>=0.4.7",
    "soundfile>=0.12.1",
    "pyannote.audio>=3.3.0",
    "torch>=2.3.0",
    "torchaudio>=2.3.0",
    "pyperclip>=1.9.0",
    "pynput>=1.7.7",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "tomli>=2.0.1",
    "pydantic>=2.7.0",
    "rapidfuzz>=3.9.0",
    "sentence-transformers>=3.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]
personalize = [
    "dspy-ai>=2.4.0",
]

[project.scripts]
flow = "flow.__main__:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["flow"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
